Create triangle.rb
A program in Ruby to find the maximum sum of numbers in a single path of an equilateral triangle, of base to height containing n X n rows and columns of numbers, respectively. A path-sum is calculated based on a number set comprised of the path through the triangle, starting at the top and moving to adjacent numbers on the row below until reaching the base.